move client_response into its own file
